requirements.txt
gdown
unzip
loguru
Pillow
tqdm
wandb
ultralytics
opencv-contrib-python==4.5.5.62
opencv-python-headless==4.5.3.56